All basic stuff completed. Missing tests and albums

+++ db/migrate/20101022022020_create_events.rb
@@ -6,8 +6,8 @@ class CreateEvents < ActiveRecord::Migration
       t.string :location
       t.datetime :start_time
       t.datetime :end_time
-      t.integer :latitude
-      t.integer :longitude
+      t.float :latitude
+      t.float :longitude
       t.references :user
 
       t.timestamps
